backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.17,<3.18
djangorestframework-simplejwt>=5.5,<5.6
psycopg[binary]>=3.2,<3.3
gunicorn>=23,<24
whitenoise>=6.9,<7
django-cors-headers>=4.7,<5
pywebpush>=2.1,<3
